combine the two files into a single one
